+++ app/src/main/java/com/techcourse/MyWebAppInitializer.java
@@ -1,0 +1,6 @@
+package com.techcourse;
+
+import com.interface21.webmvc.servlet.AbstractDispatcherServletInitializer;
+
+public class MyWebAppInitializer extends AbstractDispatcherServletInitializer {
+}
