+++ mvc/src/main/java/com/interface21/webmvc/servlet/AbstractDispatcherServletInitializer.java
@@ -1,7 +1,6 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
 import com.interface21.web.WebApplicationInitializer;
-import com.interface21.webmvc.servlet.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,9 +9,9 @@ import org.slf4j.LoggerFactory;
  * Base class for {@link WebApplicationInitializer} implementations that register a {@link DispatcherServlet} in the
  * servlet context.
  */
-public class DispatcherServletInitializer implements WebApplicationInitializer {
+public abstract class AbstractDispatcherServletInitializer implements WebApplicationInitializer {
 
-    private static final Logger log = LoggerFactory.getLogger(DispatcherServletInitializer.class);
+    private static final Logger log = LoggerFactory.getLogger(AbstractDispatcherServletInitializer.class);
 
     private static final String DEFAULT_SERVLET_NAME = "dispatcher";
